# `<regex>`: Speed up skip optimization for default `regex_traits` in `collate` mode