@StephanTLavavej 2025-08-15
I'm mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.